Allow sharing of instances with collaborators.
Create a share instance page with a box to enter an email address.
If we know the email address already, add it to the instance.
For unrecognised emails, create a user with that email and add to the instance.
If the email address is short enough, it'll be used as the username, otherwise,
we use a portion of a hash of it instead.

In either case, redirect the sharer to a confirmation page on success.

The new user gets sent an email inviting them to contribute to the sayit
instance, which a link to follow at which they can set a password and then
be redirected to the instance home page to join in with curating it.